requirements.txt
python-telegram-bot==20.4
python-magic
clamd
requests
